Add utility functions for button entity unique ID generation
- Implemented `build_speedtest_button_unique_id` to create a unique ID for the Run Speedtest button.
- Implemented `build_reset_button_unique_id` to create a unique ID for the Reset Gateway button.
- Added module docstring and defined `__all__` for public API exposure.
