+++ custom_components/diagnostics.py
