+++ custom_components/speedtest.py
