+++ hacs.json
@@ -4,7 +4,6 @@
   "homeassistant": "2024.2.0",
   "render_readme": true,
   "zip_release": true,
-  "content_in_root": false,
-  "logo": "custom_components/unifi_gateway_refactored/logo.svg",
-  "icon": "custom_components/unifi_gateway_refactored/icon.svg"
+  "logo": "custom_components/logo.svg",
+  "icon": "custom_components/icon.svg"
 }
